Fix error state in swap and move some things around
