+++ src/pages/Send/index.tsx
@@ -35,16 +35,11 @@ import TokenLogo from '../../components/TokenLogo'
 import { DEFAULT_DEADLINE_FROM_NOW, INITIAL_ALLOWED_SLIPPAGE, MIN_ETH } from '../../constants'
 import { useV1TradeLinkIfBetter } from '../../data/V1'
 import { useTokenContract, useWeb3React } from '../../hooks'
+import { useApproveCallback } from '../../hooks/useApproveCallback'
+import { useSwapCallback } from '../../hooks/useSwapCallback'
 import { useUserAdvanced, useWalletModalToggle } from '../../state/application/hooks'
 import { Field } from '../../state/swap/actions'
-import {
-  useApproveCallback,
-  useDefaultsFromURL,
-  useDerivedSwapInfo,
-  useSwapActionHandlers,
-  useSwapCallback,
-  useSwapState
-} from '../../state/swap/hooks'
+import { useDefaultsFromURL, useDerivedSwapInfo, useSwapActionHandlers, useSwapState } from '../../state/swap/hooks'
 import { useHasPendingApproval, useTransactionAdder } from '../../state/transactions/hooks'
 import { useAllTokenBalancesTreatingWETHasETH } from '../../state/wallet/hooks'
 import { CursorPointer, TYPE } from '../../theme'
@@ -668,7 +663,7 @@ export default function Send({ history, location: { search } }: RouteComponentPr
             }}
             id="send-button"
             disabled={!isValid}
-            error={severity > 2}
+            error={isValid && severity > 2}
           >
             <Text fontSize={20} fontWeight={500}>
               {error || recipientError || `Send${severity > 2 ? ' Anyway' : ''}`}
